separate clobj
